+++ app/lib/active_markdown.rb
@@ -7,6 +7,7 @@ class ActiveMarkdown
   INITIAL_VERSION = '0.3.16'
 
   def call(value:, version: nil, initial_version: nil)
+    return value if value.blank?
     return demarked_value(value) if version.nil?
     return demarked_value(value) if Gem::Version.new(version) < Gem::Version.new(initial_version || INITIAL_VERSION)