+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Modified
 - toggling features with energy
 
+### Fixed
+- markdown transformation
+
 ## [0.3.20] - 2025-11-29
 ### Added
 - rendering type and recall cost for domain cards for DH characters